requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Platform communication
python-socketio>=5.10.0

# Utilities
python-dotenv>=1.0.0
